add some ui to install.sh

+++ install.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+. scripts/ui.sh
+
 DEBUG="true"
 
 CONF_FILE=".git-sync.toml"
@@ -6,15 +8,21 @@ CONF_PATH="$HOME/.git-sync.toml"
 
 # debug
 if [ "$DEBUG" = "true" ];then
-	rm ~/.git-sync.toml
+	if [ -f "$CONF_PATH" ]; then
+		rm ~/.git-sync.toml
+	fi
 fi
 
 # test if git is installed
+dividing_line
 echo "check if git is installed..."
 if command -v git >/dev/null 2>&1;then
+	show_ok
 	echo "git is installed"
+
 else
-	echo "can not find the git"
+	show_warn
+	echo "c not find the git"
 
 	echo "if you want us to install git for you?(Y/n):"
 	read INPUT
@@ -27,10 +35,11 @@ else
 		unset INPUT
 		exit 1
 	fi
+	show_ok
 	echo "installed git successfully"
 fi
 
-echo ""
+dividing_line
 echo "installing the configuration file..."
 # test if the git-sync.toml
 if [  -f "$CONF_PATH" ];then
@@ -38,12 +47,16 @@ if [  -f "$CONF_PATH" ];then
 else
 	echo "creating the configuration file in $CONF_PATH"
 	cp $CONF_FILE $CONF_PATH
+	if [ -f "$CONF_PATH" ]; then 
+		show_ok
+		echo "installed the $HOME/.git-sync.toml successfully"
+	else 
+		show_err
+		echo "something went wrong when we are trying to write to $CONF_PATH"
+	fi
 fi
-
-
-echo ""
+dividing_line
 echo "we need to configure your git server account"
-
 
 # test the server 22 port is open
 SERVER_HOST=""
@@ -57,10 +70,12 @@ do
 	fi
 
 	if  timeout 2s nc -z $SERVER_HOST $SERVER_PORT 2>/dev/null ; then
+		show_ok
 		echo "you host is valid"
 		break;
 	else
 		echo $SERVER_HOST $SERVER_PORT
+		show_err
 		echo "the host is not valid"
 		echo "please re-input your host and port"
 		echo ""
@@ -68,11 +83,17 @@ do
 done
 
 # set the user
+dividing_line
 USER_NAME=""
 echo "please input your user to login the remote server"
-echo "maybe the current user?$USER"
+echo "maybe the current user?$USER if you press enter"
 read -p "username:" USER_NAME
+if [ "$USER_NAME" = "" ]; then 
+	USER_NAME="$USER"
+fi
 
+
+dividing_line
 TOKEN_PATH=""
 # set the ssh token path
 while [ 1 ];
@@ -92,6 +113,10 @@ sed -e "s,host.*,host = \"$SERVER_HOST\"," -i $HOME/.git-sync.toml
 sed -e "s,port.*,port = $SERVER_PORT," -i $HOME/.git-sync.toml
 sed -e "s,username.*,username = \"$USER_NAME\"," -i $HOME/.git-sync.toml
 sed -e "s,token-path.*,token-path = \"$TOKEN_PATH\"," -i $HOME/.git-sync.toml
+
+dividing_line
+show_ok
+echo "your configuration file is generated successfully"
 
 # some deprecated code to congure the github
